Reset file pointer if truncated
Previously, if the log file was truncated, the code did not reset the read position, potentially resulting in errors. This change checks the file's size, resets the read position if the file size is smaller than the last read position, and updates the tracking of the last read position accordingly.